requirements.txt
# Core
python-dotenv
requests
tqdm
pyyaml

# LLM & NLP
openai>=1.0.0
anthropic
groq
google-generativeai
transformers>=4.40.0
sentence-transformers
sentencepiece
torch>=2.0.0
tiktoken

# Text Processing
nltk
spacy
beautifulsoup4
thefuzz[speedup]

# Data Handling
pandas
numpy

# Backend / API
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
passlib[argon2]
python-jose[cryptography]

# File parsing
python-docx
pypdf
llama-parse
llama-index-core

# Orchestration / Agents
langgraph
pydantic
langchain>=0.2.14
langchain-openai>=0.1.6
langchain-mistralai>=0.1.0
langchain-core 
Jinja2

# Utils
rich
colorama
loguru

# Tests
pytest
pytest-cov

# Dev tools
black
flake8